refactor: extract HCE implementation into dedicated crate
