+++ hce/src/eval_space.rs
@@ -1,5 +1,5 @@
 use super::HCEConfig;
-use crate::hce::context::EvalContext;
+use crate::context::EvalContext;
 use cozy_chess::Color;
 use utils::{majors, minors};
