+++ hce/src/pawn_cache.rs
